Revert c846590..a5fa38e
This rolls back to commit c846590a46440949a5c1b505ea9af43ef7de7462.

+++ src45x/gcc/gcc/testsuite/gcc.target/mips/octeon-bbit-2.c
@@ -5,19 +5,19 @@
 /* { dg-final { scan-assembler "\tbnel\t" } } */
 /* { dg-final { scan-assembler-not "\tbne\t" } } */
 
-NOMIPS16 long int
-f (long int n, long int i)
+NOMIPS16 int
+f (int n, int i)
 {
-  long int s = 0;
+  int s = 0;
   for (; i & 1; i++)
     s += i;
   return s;
 }
 
-NOMIPS16 long int
-g (long int n, long int i)
+NOMIPS16 int
+g (int n, int i)
 {
-  long int s = 0;
+  int s = 0;
   for (i = 0; i < n; i++)
     s += i;
   return s;
